HPCC-29610 Fix broken file spray when no DropZoneRestriction
1. When spraying a file with no DropZoneRestriction, the
getAndValidateDropZone() may not return a dropzone plane. The
readAndCheckSpraySourceReq() should validate the return and read the
plane name if there is the return.
2. In the checkDZScopeAccessAndSetSpraySourceDFUFileSpec(), validate
DZPathScopePermissions only if the dropzone plane name is found in
the readAndCheckSpraySourceReq().
3. Create throwOrLogDropZoneLookUpError(): if not bare metal or
DropZoneRestriction enabled, throw an exception; otherwise, log the
error.
4. Create findDropZonePlaneName(): if a plane name is not found based
on dropzone host and path, call throwOrLogDropZoneLookUpError().
5. In ESP FileSpray service, call the findDropZonePlaneName() to find
a dropzone plane name. Validate DZPathScopePermissions only if the
plane name is found.
6. In getDZPathScopePermissions(), if a dropzone plane is not found
based on dropzone host and path, call throwOrLogDropZoneLookUpError().

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>