HPCC-29708 Forward compat. of dropzone groups
HPCC-29553 introduced Dali groups for dropzones, so that
logical files could map to named groups.

However, prior to this change (<= 9.0) each client populated planes
based on the existing Dali groups.
If an environment had the HPCC-29553 changes, but was then rolled
back to an earlier build, these dropzone groups were treated as
standard groups (not dropzone groups).
The consequence in the old code was that when the client code
went to add a plane for a dropzone (based on the environment), it
found there was a pre-existing GroupInformation entry and abandoned
the attempt.
As a side-effect of that, dropzone planes were not found, causing
a 'DropZone not found for host' error.

Fix by ensuring the existing <=9.0 client code instantiates the
GroupInformation instance if adding a dropzone group that clashes
and is of unknown type.

NB: This change is not needed prior to 9.0 because prior to 9.0,
dropzones were not found via planes, but via code that walked
the legacy environment.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>