HPCC-29303 Fix localhost not in dropzone error in k8s
In k8s mounted dropzones interact with EclWatch (at least at
the moment for historical reasons) with a hostname.
The service returns "localhost" for planes without hosts,
and other serivce calls to esp send "localhost" when the
content of a dropzone is interacted with.

Avoid treating this "localhost" usage as invalid.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>