HPCC-29622 Minimize number of external roxie certs per cluster
External facing certificates identify the services not the pods,
so one certificate per cluster with multiple DNS entries should
suffice.

Also ZeroSSL takes a long time to generate certificates and
LetsEncrypt has limitations so there are issues if we don't
keep the number of external certificates to the minimum
necessary.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>