HPCC-29970 Check case-insensitive owner ID in LF/WU
The owner ID of a logical file or workunit is based on user's
login name which can be case-insensitive. The code for the ID
verification should allow the case-insensitive ID as we do in
CDaliWorkUnitFactory.getWorkUnitsByOwner().
1. Change the setFileProtectTree(), getNamedPropTree(),
getAttrQueryStr(), and related code to allow case-
insensitive owner ID.
2. Allow case-insensitive owner ID when validating ECL WU
access by ownership.
3. In the IDFUWorkUnitFactory.getWorkUnitsByOwner(), allow
case-insensitive owner ID and wildcard match (the same as
in CDaliWorkUnitFactory.getWorkUnitsByOwner()).

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>