HPCC-30310 Allow esdl ecl command to output to stdout
Update esdl ecl command to allow output to stdout by making the outputDir
parameter optional. When using stdout, restrict the use of options to those
combinations that result in a single file output.

Disabled verbose output in the call to EsdlCmdHelper::convertECMtoESXDL to
avoid printing extra text to stdout.

Update esdl regression test with cases for esdl ecl, though the suite needs
independent updates to bring tests of the other commands in line with
recent platform changes. This will be handled in the ticket HPCC-30699.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>