HPCC-30704 Prevent scale down evict
It is generally not safe for HPCC pods to be evicted based on
scale down events because most are stateful (particularly those
that have been dynamically launched as jobs).
For now, prevent all pods from being deemed suitable for eviction
by k8s on scale-down.
We may later allow some that have replicas and aren't stateful
or 'cheap' to evict and move and handle graceful termination
to be safe, but it will then be better to use PodDisruptionBudget's

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>