HPCC-26634 Check query status only on 'can publish queries' roxies
For containerized, the existing WUListQueries tries to check
suspended queries for all roxies. The existing WUListQueries
returns a query not found error if there is a non-queries roxie.
For bare metal, the existing WUListQueries tries to get suspended
queries for every target which contains a RoxieCluster process.
The same error occurs for the targets which cannot publish queries.
The WUListQueries should not check suspended queries for the
roxies which cannot publish queries.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>