HPCC-30873 Add Ensure option to file services and dfuplus
The ability to copy a logical file without copying the physical
files was added if they are already present on the target.
Before this PR, that option is only available at the ESP service
level via a SOAP call. Now, this option is added to fileservices
and the dfuplus command line.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>