HPCC-30854 Ensure partitioning of KJ parts between workers in k8s
When mapping index parts to workers in containerized, treat all as
if locally available, such that they will be round-robined into
worker buckets. This will cause (when the default remoteKeyedLookup
is true), workers to directly access a partition of the index parts
and farm out request of the rest to other workers.
The net result is to increase cache locality (both jhtree and within
the Linux page cache).

Also make some detailed logging conditional on logging level.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>