HPCC-30848 Read remote storage settings into ESP
This PR fixes the error of the 'remote storage not found' when
copying a file from cloud to bare metal. The fix includes: 1.
copying the remote storage settings from environment.xml to the
esp.xml using the esp.xsl; 2. reading the settings from the
esp.xml to the GlobalConfig when a bare metal ESP is started.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>