HPCC-30937 Update readCost, writeCost, numDiskReads and numDiskWrites periodically
Previously, readCost, writeCost, numDiskReads and numDiskWrites was
updated to the workunit at the end of the activity.  This meant that
users were unable to view these stats while the activity was running.
Furthermore, as these costs were calculated at the activity, the
guillotine was also not always effective until the end of an activity
(i.e activities that used a lot of disk access to exceed the guillotine
threshold substantially before being terminated.)

This commit updates readCost, writeCost, numDiskReads and numDiskWrites
periodically (with every serialization of the stats to master).  The
updates to file properties will still be updated at the end of the
activity to avoid creating extra load on dali.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>