HPCC-30998 Fix post-mortem files being attached to wrong workunit.
With multiJobLinger on (now the default), post-mortem files were
being attached to the 1st workunit than ran on a started Thor
instance, rather than the one that had most recently ran and
caused the reports.

Fix by recording the current wuid to a local temp file, for
check_executes.sh to use when attaching post-mortem files.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>