HPCC-31038 Add option for removing 'lost' file from super file
In the existing code, a 'lost' file found by XRef can't be deleted
if that file is a sub of a super file without manually being removed
from the super file. With this option is set inside delete request,
the file will be automatically removed from its owning super files
before the 'lost' file is deleted.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>