HPCC-31028 Use cost_type internally for consistency
Costs have been represented as double in some places in the code
and cost_type in other places.  The use of both double and
cost_type internally for costs has made the code more difficult
to understand and maintain.  It has lead to some bugs.

This commit makes all internal representation of cost in the code
as cost_type.  The cost_type is only converted to double when
returning costs externally via ESP services.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>