HPCC-31035 Move ctx stats merging to caller from CKeyLevelManager mergeStats
Prior to this change, it wasn't possible to share a IContextLogger
between different CKeyLevelManager.  The reason was that a
mergeStats call needed to made of each of the CKeyLevelManagers
but each call to the CKeyLevelManager::mergeStat would have the
same jhtree stats being merged.

By removing the IContextLogger merge from CKeyLevelManager::mergeStats
it is possible for the same IContextLogger to be shared between
different CKeyLevelManagers.  This change required, the caller to
merge the stats from the IContextLogger.

This change make the IndexRead activity jhtree stats change
simpler and more efficient as it can use the same IContextLogger from
the different file parts.)

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>