HPCC-23064 Improve error messages when no access to ECL WUs
When access denied to ECL WUs, the WsWorkunits should report which
access feature (OwnWorkunitsAccess or OthersWorkunitsAccess) is
denied so that an administrator may know how to change user's
access permission.
The code changes are:
1. add the code to report which access feature is validated in the
validateWsWorkunitAccess() and validateWsWorkunitAccessByOwnerId().
If the validation fails, the feature name can be used inside error
messages.
2. Change the validateWsWorkunitAccess() function to 2
validateWsWorkunitAccess() functions: one for IConstWorkUnit input
and one for wuid input. The 1st one is called by the 2nd one and
other functions.
3. In the functions below, if the WU access validation fails, add
the name of the access feature into the error message:
  ensureWsWorkunitAccess()
  ensureWsWorkunitAccessByOwnerId()
  ensureWsCreateWorkunitAccess()
  doAction().
4. In the doWUQueryByFile(), call the validateWsWorkunitAccess()
to validate WU access. Change the error message if the validation
fails. Also clean the code.
5. Simplify the onWUQuery by switching validateWsWorkunitAccess()
call to the ensureWsWorkunitAccess() call.
6. In the onStartUpload(), when importing a WU ZAP report, change
the error message if the user does not have the access permission.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>