HPCC-31122 Remove ZAP 'Log Access plug-in' warning for bare-metal
When creating a ZAP report, if Remote Log Access plug-in is not
available, ECLWatch shows a warning: "Warning: may not be able to
include log file because Remote Log Access plug-in is not available!".
Before this PR, the warning shows for both cloud and bare-metal. In
this PR, the warning is removed for bare-metal since the Remote Log
Access interface is not used in bare-metal at the moment.

Revise based on review:
1. Revise the warning message.
2. If Containerized, skip the code to report thor processes.
3. Add comments about reporting thor processes for bare-metal.

Signed-off-by: wangkx <kevin.wang@lexisnexis.com>