HPCC-30996 Fix issues with roxie copying from remote storage source
Roxie deployment (via esp) creates new logical file entries in the
local Dali and annotates it with cloneFrom* info, so that at
runtime, Roxie can read remotely, and copy file parts
asynchronously in the background to the local placements.
This did not work where the remote file was from a non-standard
topology, in particular where the remote files were from cloud,
where parts are by default stored per directly, and where storage
may be striped.
Fix by cloning the source file descriptor as a reference point,
such that the remote locations and fully use the remote file
descriptor as a source.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>