HPCC-31141 Validate global sort connect protocol
Add a const signature to the global sort connect protocol,
and validate it in the header on connect.
This is to exclude 'rogue' attempts to connect to the sort socket,
possibly from k8s health type checks.

Also remove pointless byte ordering on serialize/deserialize in
this context, where both sides are implicitly on same
OS/byte-order/version.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>