HPCC-31290 Fix Sasha Thor QMon switching issues
Automatic and manual queue switching by the Sasha QMon service
have not worked since 7.12.
Changes to the format of the queued thor job items meant it did
not find any queued items to swap.
This meant that workunits submitted with 'allowedclusters'
and 'allowautoqueueswitch', that should have automatically
switched to an idle Thor queue in the 'allowedclusters' set
when the queue they were submitted to was busy, did not.

Also fix the qmon tracing, which was supposed to trace
the current workunits in flight running on Thor instances.
That appears not to have worked well before 7.12.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>