HPCC-31160 Fix slow CSV read of super with HEADINGs.
A CSV read of a superfile with HEADING(<n>) was causing a massive
slowdown. Each worker was waiting for the previous to have
processed all header lines from all subfiles before moving
forward. The net effect was for Thor to process the file
sequentially on each worker instead of in parallel.

NB: kept old code that pended on previous worker to have
processed all subfile header lines as a failsave expert option.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>