HPCC-31286 Potential stall if out of roxiemem in lookup join
If lookup hits an out of roxie mem. exception whilst processing
rows on some slaves, it could potentially cause a stall.
It may be that it can only happen if it occurs after it the
row processor has received all chunks.

Abort the broadcast when handling the exception in the row
processor.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>