HPCC-31341 Fix DFU key copy issue if copying to different sized cluster
If copying a key, wrap will be used, and the target part size must
match the source. The target number of parts was being set into
numPartsOverride, which was then picked up by the dfu job to
ensure the target file was created with the same number of parts.
However, the same setting is used as a request via spray/copy
services and is not permitted to be used on a key copy.

If the DFU workunit was resubmitted, the persisted numPartsOverride
was detected and a 'DestinationNumPartOverride is provided but
getWrap is true' error was fired.

Avoid setting numPartsOverride, instead set numParts,
thus avoiding the error if the DFU workunit is resubmitted.
Also allow NumPartsOverride as long as it matches source # parts.
And improve the error message.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>