HPCC-29812 Mitigate rename failures on k8s with nfs backed storage.
Occassionally, renaming a physical file that is on a mount backed
by Azure Blob storage over NFS fails. That is ::rename returns
-1 indicating an error. However, on inspection immediately after
the failure, the rename has actually occurred, i.e. the src file
no longer exists, and the dst does.

Add a mechanism and configuration option to allow CFile::rename
to continue after failure if it can establish that the source
filename has actually been renamed to the target filename.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>