HPCC-31394 WsSMC send roxie control cmd to ssl port if available
Rather than default to port 9876, use the roxie farmer ports configured
for sending control messages to. Prefer an ssl port if available.

Update the smart pointer factory and ctors where required to accept a tls
config from bare-metal configurations. These configurations using transitional
APIs don't support all the properties and sync features of the containerized
versions.

A separate ticket would be required to determine if and how to add the full
complement of tls related configuration, behavior and secrets support to bare
metal that is currently in place in containerized code.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>