HPCC-31474 Refactor dafilesrv handling to cope with SSL/non-blocking
1) Switch accepted sockets to non-blocking
2) On notifyConsume available bytes, including partial
writes of leading message size if necessary.
Return from select handler asap.
3) On full message, handover to threadpool

This works by trying to read expected sizes, and via
non-blocking returning immediately if not available.

Previously, there was a convoluted set of stages, and some
problems due to the wait it depended on 'avail_read', which did
not work as expected when using SSL sockets (it caused hangs).

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>