HPCC-31468 Prevent propagation of AKS remote file meta info
When a k8s logical file's meta data is retrieved, it is
supplemented with info. about the remote plane characteristics
the remote file resides on. This is used by the clients DFS to know
how to map file into in lookup calls to the k8s dafilesrv service.

The logical file meta data as a whole can be stored as extra
meta data by the roxie file deploy mechanism, which is then used
as a source when Roxie is starting, to copy locally.

If a secondary Roxie uses a Roxie which copied from a k8s, it
will also see and use this remote plane info spuriously, and
then use it to manipulate the group/file paths, to point to the
original k8s source.

Suppress this info. being propagated.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>