HPCC-28452 ECL Watch fix "Session is locked" messages
Remove any client code that would write a "Status" cookie when either
the session timeout monitor fires, or the user manually clicks the
"Lock" link.

Also, though not strictly related, this PR includes a fix for one
of the rewrite rules for the lightweight web server config
(used for local dev of ecl watch).

Signed-off-by: Jeremy Clements <79224539+jeclrsg@users.noreply.github.com>