HPCC-31652 Guard against calculateSkew causing attach to fail
When attaching a file to DFS it performs a skew calculation,
which may check physical file parts if there are missing size
attributes. If a physical file is missing during this process,
it will cause attach() to fail.
A pull Roxie attaches a logical file without the physical files
being present, if size attributes are missing it will fail.
Protect against this failure in the skew calculation
(catch exception, log and continue).

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>