HPCC-31452 Fix coverity-flagged move semantics defect
Use result.get() to force the ternary operator to see its args as const char*.

Defect appeared to be a false positive until Gavin pointed out that the
StringAttr conversion constructor for const char* may be causing the ternary
operator to use StringAttr temporaries. Coverity likely saw that as an
opportunity for optimization with move.

However, the ambiguous multiple paths for implicit type conversion is the
true problem which may have raised an error in the future.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>