HPCC-31760 Protect against delta save failures
If there are temporary failures (e.g. failures to write to disk),
then protect against the delta writer thread from exiting.
Retry all disk actions.
NB: the initial inspiration for this was that there was a failure
due to Dali running out of file handles. Which itself is probably
caused by the flood of connections from AKS VM's to Dali.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>