HPCC-31957 Remove unnecessary exclusive lock on source file
During an XML spray/despray filecopy was unecessarily obtaining
an exclusive write lock on the source file attributes.
If the file was being read elsewhere, this could cause a
deadlock whilst it waited for the exclusive lock.
(in one case it was observed that 2 jobs were started performing
a despray of the same source file, and they deadlocked each
other)

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>