HPCC-31844 Fix dll cache in containerized setup
The manager's query dll cache was not being initialized in
containerized, which meant that it could cache an indeterminate
amount of query dlls. However, the workers did initialize their
cache, meaning that their copies would be deleted as expected.
The net result was that the manager would consider it cached
and not resend a dll, causing the slave to fail since its copy
had been deleted.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>