HPCC-32146 Fix logic waiting for workunit complete to spot hard fail
The workunit k8s launch code, waits on the state of the workunit
after it's lauched. Once it responds, which includes if it disappears
and the queuing agent (thoragent) marks it as fails, it scans the
workuinit for exceptions and throws one if it exists.
However, if there were some exceptions, but no 'thormasterexception'
reply exception, then the logic skipped the follow on test to
see if the workunit had been marked failed, and continued.

This results in either jobs appearing to run indefinitely, or
spuriously marking themselves as complete

NB: reproduced by deleting the thormanager pod abruptly (--force=true)

There also some changes her which wrap the deleteResource commands
in case they throw exception, because we do not want those exceptions
to cause job failures in themselves, or mask the original exception.
Log as operator warnings instead.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>