HPCC-32241 Temp size for sort to use disk size and track actual graph temp disk usage
SizePeakTempDisk for graphs was previous calculated by summing the each
activity's active temp file size.  However, as this was done at internals,
spikes in temp file sizes between intervals were not recorded in the
SizePeakTempDisk stats.  This change uses a TempFileSize tracker at the
graph level to track the actual peak/active temp file size for the graph.

Sort activities was using the size of the uncompressed data as the size
of the temp file.  This change uses the actual size of temp file for sort
activity's peak/active temp file size

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>