HPCC-32313 Limit lookup broadcast queue size (preventing mem build up)
As a smart join is failing over to a local smart join and stopping
the broadcaster, an excessive number of inflight packets can be
received and queued on the broadcaster queue. If this is unbound
it can cause a large amount of heap memory to be consumed.

Also:
1) add logic into the broadcaster so that receivers send back
notification that they are stopping to the sender, so the sender
can stop as quickly as possible.
2) Prevent adding packages to the broadcaster thread if it
can be deemed that they do not need broadcasting anywhere else.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>