HPCC-32421 Suppress foreign groups
Prevent foreign groups from being published to Dali.
The code used to try to find a matching group (for naming purposes),
and then create a unique one. What made it worse, was that if
the foreign group was from an AKS system, it would have ports
associated with it, and the published groups never have ports,
consequently it never matched and always added a new unique (anon)
group.
This caused a gradual build up of new groups (under /Group), which
cause delays for everything that looked up a group elsewhere.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>