HPCC-32554 Fix WorkUnitWaiter so it can be used to spot abort correctly
Queued k8s jobs were not being aborted. The WorkUnitWaiter was
used to monitor the state, but it was not told to spot abort conditions
(SubscribeOptionAbort), and the implementation didn't correctly
detect the aborting condition.
Fix and ensure waitForWorkUnit returns 'aborting'.
Also ensure that Thor doesn't overwrite aborting (or aborted) state
with failed.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>