HPCC-32661 Sort k8s workers to avoid storage stripe clumping
Previously the worker order was indeterminate, they would
register on a 1st come 1st served basis.
That could cause workers that striped to a particular
storage account to clump.
Sort by pod/container name, to reduce the number that
would modulate to the same storage account from clumping on
the same pod.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>