HPCC-32307 Fix inappropriate setting of dirPerPart flag in roxie cloneSubFile
This bug would cause 1-way files copied to an k8s roxie to
inappropriately have the 'dirPerPart' flag set. If the physical files
were copied via other means, then it caused a mismatch between the
meta data and physical layout, causing the physical parts of 1-way
files not to be found leading to queries being suspended.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>