HPCC-32822 Fix MP protocol error logged indefinitely in loop
The detection of a MP packet with an invalid header threw
an exception that was logged, but did not close the socket.
When the client closed the socket, the MP server was notified
but because it had nothing left to read of the [bad] header,
it saw the invalid protocol error again and rethrew the
same error.
The epoll handler loop continuously notified the handler
of the close event (because it did nothing with it), and
the protocol error was continuosly output.

The socket should be close on this and any other exception.

In case these events are too frequent, add a timer to log
the protocol errors less frequently, and log the header
bytes received to help diagnose what the source of these
'rogue' connections are.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>