HPCC-32803 In file ops, update source file props without locking source file
This fixes the issue of some file operations failing because the
source file was locked when the operation was taking place. Source
files may be locked if it is being processed elsewhere.  Prevously,
the source file was being locked to allow the source file's disk
read count and read costs properties to be updated.  This issue is
fixed by making use of CFileAttrLock which makes it possible to
update file properties without requiring the file to be locked.

Changes:
- add addReadCost and addNumDiskRead to IDistributedFile interface
- Implementation of addReadCost and addNumDiskRead makes use of
CFileAttrLock to lock the file properties and make the necessary
updates
- Make FileSprayer::updateTargetProperties update only the target
properties (it was also previously updating the source properties)
- Create FileSprayer::updateSourceProperties to update the source
properties and make it use IDistributedFile::addReadCost and
IDistributedFile::addNumDiskRead to make the necessary updates,
removing the need to lock the IDistributedFile.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>