HPCC-32877 Fix abort in esp when accessing secrets from a vault
The vault secret code uses httplib to request entries
from a vault.  The default compilation mode is to use select() rather
than poll(), but select() has a limit of 1024 sockets.  That means
that if a process has a large number of sockets (or files?) open
the select based code will fail.

This commit defines CPPHTTPLIB_USE_POLL which enables the use of poll()

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>