HPCC-32919 Fix backward compatibility problems with old roxie workunits
When some of the new regex optimizations were added the new functions were not added at the end of the class.  That means that existing workunits will no longer be compatible.

Fix the ordering so that backward compatibility is retained.  This will cause backward incompatibility within any 9.8.x roxie deployments.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>