HPCC-32947 Capture and report lookahead timings for project
The modification records time spent in the strands as lookahead
timing.  This is necessary to ensure the local execute timings
for project and upstream activities are correct.

Ideally, this lookahead timings would be recorded a separate
statistic such as "background activity".  However, a new
statistic will not be created at this time to avoid complicating
the statistics.

Signed-off-by: Shamser Ahmed <shamser.ahmed@lexisnexis.com>