HPCC-32948 Make ws_logaccess WSDL match exception return behavior
The service returns exceptions inline, but the WSDL does not show the
exceptions array in the method response elements.

Fix the interface defn to ensure the WSDL shows exceptions are returned
inline.

Note these caveats:

- This change cannot be versioned and will retroactively change
the WSDLs for all service interface versions. The alternative is that a
current exception response would not validate against the WSDL.
- Even with the fix the response will not match the WSDL in some cases.
Transactions sent with "Content Type: application/x-www-form-urlencoded"
(most of the requests from ECLWatch) return inline exception responses
missing the root "ESPresponse" element. The root of the response is the
"Exceptions" element.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>