HPCC-31872 Accept alternative domains in certificate config
Accept an `alternativeDomains` configuration property in the certificates
issuers section. It is an array of domains other than `domain` that the
certificate will be valid for.

These values will be added to the helm manifest's Certificate resource's
`dnsNames` property, which corresponds to entries in the "Subject Alternative
Name" extension of the X.509 certificate.

The `domain` value is used in part to derive the `commonName`, and it is still
added to the list of `dnsNames` as before. Any ESP and roxie service names are
prepended to any of the `domain` and `alternativeDomains` values before adding
to the `dnsNames` list.

Signed-off-by: Terrence Asselin <terrence.asselin@lexisnexisrisk.com>