HPCC-33022 Fix pathological dali load time
JPTree (used by Dali SDS) could perform horrendously slowly when
dealing with a large list of child nodes sharing a common
prefix. This was due to how the hashing function was used
(no seed number provided). With many entries having the same
prefix, it caused a high number of hash collisions, resulting
in excessive scanning when adding new members.
This particularly affected the /GeneratedDlls section of Dali.
After fixing the issue, the time to load this section dropped
from over 4700 seconds to 17 seconds.

Also add a 'loadxml' utility to daliadmin, that loads and
times loading an xml file, and traces the amount of memory it
consumes.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>