HPCC-33100 Fix data loss, streaming from dafilesrv and network failure.
If there is a broken socket whilst streaming from dafilesrv, the
underlying mechanism will dispose of the socket and reconnect
and retry the command. If the server side remained running and did
not receive notification of the closed socket, the client can
reconnect to the same server and the retry will cause the next
chunk of data in the stream to be retrieved, thereby losing the
earlier one that hit the socket issues.
Fix by avoiding the reconnect/retry mechanism at this level,
instead, on failure during streaming, establish a new stream
at the valid position using the serialized cursor.

Signed-off-by: Jake Smith <jake.smith@lexisnexisrisk.com>